Introduce Constable.ConnTestHelper
It is considered bad practice to define functions in the quoted block.
It is easier to test, modify and reuse functions if you import a module
instead.
