+++ test/support/conn_case.ex
@@ -21,6 +21,7 @@ defmodule Constable.ConnCase do
       # Import conveniences for testing with connections
       use Phoenix.ConnTest
       use Constable.Web, :view
+      import Constable.ConnTestHelper
 
       # Alias the data repository and import query/model functions
       alias Constable.Repo
@@ -33,34 +34,6 @@ defmodule Constable.ConnCase do
 
       # The default endpoint for testing
       @endpoint Constable.Endpoint
-
-      def authenticate(user \\ create(:user)) do
-        conn = conn()
-        |> put_req_header("accept", "application/json")
-        |> put_req_header("authorization", user.token)
-        %{conn: conn, user: user}
-      end
-
-      defp fetch_json_ids(key, conn, status \\ 200) do
-        records = json_response(conn, status)[key]
-        Enum.map(records, fn(json) ->
-          Map.get(json, "id")
-        end)
-      end
-
-      defmacro render_json(template, assigns) do
-        view = Module.get_attribute(__CALLER__.module, :view)
-        quote do
-          render_json(unquote(template), unquote(view), unquote(assigns))
-        end
-      end
-      def render_json(template, view, assigns) do
-        view.render(template, assigns) |> format_json
-      end
-
-      defp format_json(data) do
-        data |> Poison.encode! |> Poison.decode!
-      end
     end
   end
